+++ core/src/main/java/spritz/Spritz.java
@@ -1,6 +1,5 @@
 package spritz;
 
-import javax.annotation.Nonnull;
 import org.realityforge.braincheck.BrainCheckConfig;
 
 /**
@@ -33,16 +32,6 @@ public final class Spritz
   }
 
   /**
-   * Return true if uncaught error handlers are enabled.
-   *
-   * @return true if uncaught error handlers are enabled, false otherwise.
-   */
-  public static boolean areUncaughtErrorHandlersEnabled()
-  {
-    return SpritzConfig.areUncaughtErrorHandlersEnabled();
-  }
-
-  /**
    * Return true if invariants will be checked.
    *
    * @return true if invariants will be checked.
@@ -60,52 +49,5 @@ public final class Spritz
   public static boolean shouldCheckApiInvariants()
   {
     return BrainCheckConfig.checkApiInvariants();
-  }
-
-  /**
-   * Return true if active tasks will be purged if the scheduler is still running after the maximum number of rounds.
-   *
-   * @return true if active tasks will be purged if the scheduler is still running after the maximum number of rounds.
-   */
-  public static boolean purgeTasksWhenRunawayDetected()
-  {
-    return SpritzConfig.purgeTasksWhenRunawayDetected();
-  }
-
-  /**
-   * Add error handler to the list of error handlers called.
-   * The handler should not already be in the list. This method should NOT be called if
-   * {@link #areUncaughtErrorHandlersEnabled()} returns false.
-   *
-   * @param handler the error handler.
-   */
-  public static void addUncaughtErrorHandler( @Nonnull final UncaughtErrorHandler handler )
-  {
-    UncaughtErrorHandlerSupport.get().addUncaughtErrorHandler( handler );
-  }
-
-  /**
-   * Remove error handler from list of existing error handlers.
-   * The handler should already be in the list. This method should NOT be called if
-   * {@link #areUncaughtErrorHandlersEnabled()} returns false.
-   *
-   * @param handler the error handler.
-   */
-  public static void removeUncaughtErrorHandler( @Nonnull final UncaughtErrorHandler handler )
-  {
-    UncaughtErrorHandlerSupport.get().removeUncaughtErrorHandler( handler );
-  }
-
-  /**
-   * Report an uncaught error in stream.
-   *
-   * @param error the error.
-   */
-  public static void reportUncaughtError( @Nonnull final Throwable error )
-  {
-    if ( areUncaughtErrorHandlersEnabled() )
-    {
-      UncaughtErrorHandlerSupport.get().onUncaughtError( error );
-    }
   }
 }