+++ core/src/test/java/spritz/AbstractTest.java
@@ -5,6 +5,7 @@ import org.realityforge.braincheck.BrainCheckTestUtil;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Listeners;
+import zemeckis.ZemeckisTestUtil;
 import static org.testng.Assert.*;
 
 @Listeners( MessageCollector.class )
@@ -17,6 +18,7 @@ public abstract class AbstractTest
   protected void beforeTest()
   {
     BrainCheckTestUtil.resetConfig( false );
+    ZemeckisTestUtil.resetConfig( false );
     SpritzTestUtil.resetConfig( false );
     _logger.getEntries().clear();
     SpritzTestUtil.setLogger( _logger );
@@ -26,6 +28,7 @@ public abstract class AbstractTest
   protected void afterTest()
   {
     BrainCheckTestUtil.resetConfig( true );
+    ZemeckisTestUtil.resetConfig( true );
     SpritzTestUtil.resetConfig( true );
   }