+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Upgrade the `com.squareup` artifact to version `1.13.0`.
 * Upgrade the `org.realityforge.braincheck` artifact to version `1.29.0`.
 * Upgrade the `com.google.testing.compile` artifact to version `0.18-rf`.
+* Replace the `VirtualProcessorUnit`, `Scheduler`, `Cancelable` etc with facilities provided by the `org.realityforge.zemeckis:zemeckis-core` artifact.
 
 ### [v0.10](https://github.com/spritz/spritz/tree/v0.10) (2020-01-20) · [Full Changelog](https://github.com/spritz/spritz/compare/v0.09...v0.10)